Chickle-Backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
google-generativeai==0.4.1  # Gemini AI SDK
requests==2.31.0
pip install --upgrade pip